Added util function
