+++ packages/util/src/promise.ts
@@ -1,0 +1,29 @@
+/**
+ * @license
+ * Copyright 2022 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { Deferred } from "./deferred";
+
+/*
+    Rejects if the given promise doesn't resolve in timeInMS milliseconds.
+    @internal
+*/
+export function timeoutResolve<T>(promise: Promise<T>, timeInMS = 2000) {
+  const deferredPromise = new Deferred<T>();
+  setTimeout(() => deferredPromise.reject('timeout!'), timeInMS);
+  promise.then(deferredPromise.resolve, deferredPromise.reject);
+  return deferredPromise.promise;
+}
